backend/requirements.txt
fastapi
uvicorn
pyjwt
bcrypt
python-multipart
python-dotenv
sqlcipher3-wheels
